ConsumerControl: allow 10-bit keycodes
In ConsumerControlWrapper, accept a 10-bit value for the keycode.
Use uint16_t since that is the type KeyboardioHID uses.

+++ src/kaleidoscope/driver/hid/keyboardio/Keyboard.h
@@ -140,10 +140,10 @@ class ConsumerControlWrapper {
     ConsumerControl.releaseAll();
   }
 
-  void press(uint8_t code) {
+  void press(uint16_t code) {
     ConsumerControl.press(code);
   }
-  void release(uint8_t code) {
+  void release(uint16_t code) {
     ConsumerControl.release(code);
   }
 };
